pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qkd-sim"
version = "0.1.0"
description = "QKD光网络仿真：物理层噪声建模与网络层资源分配"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "networkx>=3.0",
]

[project.optional-dependencies]
rl = ["gymnasium>=0.28"]
ilp = ["pulp>=2.7"]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
